function/requirements.txt
adal==1.2.7
applicationinsights==0.11.10
argcomplete==1.12.3
azure-cli-core==2.34.1
azure-cli-telemetry==1.0.6
azure-core==1.23.0
azure-cosmos==4.2.0
azure-functions==1.2.1
azure-mgmt-core==1.3.0
bcrypt==3.2.0
bson==0.5.10
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==8.0.4
cryptography==36.0.2
distlib==0.3.4
filelock==3.6.0
Flask==1.1.2
grpcio==1.44.0
humanfriendly==10.0
idna==2.10
isodate==0.6.1
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==1.0.0
knack==0.9.0
MarkupSafe==2.1.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
oauthlib==3.2.0
packaging==21.3
paramiko==2.10.3
pipenv==2022.1.8
pkginfo==1.8.2
platformdirs==2.5.1
portalocker==1.7.1
psutil==5.9.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pymongo==3.10.1
PyNaCl==1.5.0
pyodbc==4.0.32
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.23.0
requests-oauthlib==1.3.1
six==1.16.0
tabulate==0.8.9
typing_extensions==4.1.1
urllib3==1.25.11
virtualenv==20.13.4
virtualenv-clone==0.5.7
Werkzeug==2.0.3
psycopg2-binary
